+++ 1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/CannonBall.cpp
@@ -96,5 +96,5 @@ void CannonBall::HandleCollision(IEntity* pOther)
 
 SGD::Rectangle CannonBall::GetRect() const
 {
-	return AnimationSystem::GetInstance()->GetRect(m_Timestamp, (int)m_ptPosition.x, (int)m_ptPosition.y, SGD::Size{ 1, 1 });
+	return AnimationSystem::GetInstance()->GetRect(m_Timestamp,m_ptPosition.x, m_ptPosition.y, SGD::Size{ 1, 1 });
 }
